+++ benefits/src/sdk/sdkconfiguration.py
@@ -24,9 +24,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '4.0.0'
-    gen_version: str = '2.248.6'
-    user_agent: str = 'speakeasy-sdk/python 4.0.0 2.248.6 1.0.0 wingspan_benefits'
+    sdk_version: str = '4.0.1'
+    gen_version: str = '2.252.2'
+    user_agent: str = 'speakeasy-sdk/python 4.0.1 2.252.2 1.0.0 wingspan_benefits'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> Tuple[str, Dict[str, str]]: