+++ benefits/RELEASES.md
@@ -110,3 +110,13 @@ Based on:
 - [python v4.0.0] benefits
 ### Releases
 - [PyPI v4.0.0] https://pypi.org/project/wingspan_benefits/4.0.0 - benefits
+
+## 2024-02-12 01:25:15
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.177.1 (2.252.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.0.1] benefits
+### Releases
+- [PyPI v4.0.1] https://pypi.org/project/wingspan_benefits/4.0.1 - benefits